requirements.txt
# IMPORTANT: This project requires Python 3.8-3.11
# TensorFlow does not yet support Python 3.12+ 
# Please use Python 3.11 or create a virtual environment with Python 3.11

# Core Deep Learning and ML Libraries
tensorflow>=2.10.0  # Uncomment when using Python 3.8-3.11
numpy>=1.23.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.2.0
Pillow>=9.0.0

# Web Framework
streamlit>=1.25.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data Processing
opencv-python>=4.8.0

# Dataset Download
kaggle>=1.5.0
gdown>=4.7.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0

# Image Augmentation (for GAN and advanced augmentation)
albumentations>=1.3.0

# Additional utilities
scipy>=1.10.0

# Optional: Explainable AI (install separately if needed)
# grad-cam>=1.4.0

# Optional: IoT Integration (uncomment if using physical sensors)
# paho-mqtt>=1.6.0
# pyserial>=3.5
